requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
moviepy
faiss-cpu
Pillow
numpy
git+https://github.com/openai/CLIP.git
